Move viewer store and effects to stores folder
