+++ libs/search-widget/src/core/stores/viewer-effects.ts
@@ -2,13 +2,13 @@ import type { Subscription } from 'rxjs';
 import { concatMap, filter, switchMap, tap } from 'rxjs/operators';
 import { getResource, loadEntities } from '../api';
 import type { Resource } from '@nuclia/core';
-import { resource } from '../stores/resource.store';
-import { isViewerOpen } from '../stores/modal.store';
+import { resource } from './resource.store';
+import { isViewerOpen } from './modal.store';
 import { formatQueryKey, updateQueryParams } from '../utils';
-import { canEditLabels } from '../stores/widget.store';
-import { initLabelStore } from '../stores/effects';
-import { entityGroups } from '../stores/entities.store';
-import { displayedResource } from '../stores/search.store';
+import { canEditLabels } from './widget.store';
+import { initLabelStore } from './effects';
+import { entityGroups } from './entities.store';
+import { displayedResource } from './search.store';
 import type { DisplayedResource } from '../models';
 
 const subscriptions: Subscription[] = [];
