+++ libs/search-widget/src/old-components/viewer/Viewer.svelte
@@ -8,7 +8,7 @@
     viewerStore,
     viewerState,
     selectSentence,
-  } from '../../core/old-stores/viewer.store';
+  } from '../../core/stores/viewer.store';
   import { onDestroy } from 'svelte';
   import { combineLatest, filter, map, switchMap } from 'rxjs';
   import Header from './Header.svelte';
